rename all those pwa trace requires
