+++ lighthouse-core/test/audits/byte-efficiency/render-blocking-resources-test.js
@@ -14,8 +14,8 @@ const Simulator = require('../../../lib/dependency-graph/simulator/simulator.js'
 const NetworkRequest = require('../../../lib/network-request.js');
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
