+++ lighthouse-core/test/audits/metrics/first-contentful-paint-test.js
@@ -8,8 +8,8 @@ const Audit = require('../../../audits/metrics/first-contentful-paint.js');
 const assert = require('assert');
 const options = Audit.defaultOptions;
 
-const pwaTrace = require('../../fixtures/traces/progressive-app-m60.json');
-const pwaDevtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const pwaTrace = require('../../fixtures/traces/progressive-app-m79.json');
+const pwaDevtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
