+++ lighthouse-core/test/audits/network-rtt-test.js
@@ -7,7 +7,7 @@
 
 const NetworkRTT = require('../../audits/network-rtt.js');
 
-const acceptableDevToolsLog = require('../fixtures/traces/progressive-app-m60.devtools.log.json');
+const acceptableDevToolsLog = require('../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
