+++ lighthouse-core/test/audits/network-server-latency-test.js
@@ -7,7 +7,7 @@
 
 const ServerLatency = require('../../audits/network-server-latency.js');
 
-const acceptableDevToolsLog = require('../fixtures/traces/progressive-app-m60.devtools.log.json');
+const acceptableDevToolsLog = require('../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
