+++ lighthouse-core/test/audits/screenshot-thumbnails-test.js
@@ -10,8 +10,8 @@ const path = require('path');
 const assert = require('assert');
 
 const ScreenshotThumbnailsAudit = require('../../audits/screenshot-thumbnails.js');
-const pwaTrace = require('../fixtures/traces/progressive-app-m60.json');
-const pwaDevtoolsLog = require('../fixtures/traces/progressive-app-m60.devtools.log.json');
+const pwaTrace = require('../fixtures/traces/progressive-app-m79.json');
+const pwaDevtoolsLog = require('../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
