+++ lighthouse-core/test/computed/load-simulator-test.js
@@ -8,7 +8,7 @@
 /* eslint-env jest */
 
 const assert = require('assert');
-const devtoolsLog = require('../fixtures/traces/progressive-app-m60.devtools.log.json');
+const devtoolsLog = require('../fixtures/traces/progressive-app-m79.devtools.log.json');
 const LoadSimulator = require('../../computed/load-simulator.js');
 const NetworkNode = require('../../lib/dependency-graph/network-node.js');
