+++ lighthouse-core/test/computed/metrics/estimated-input-latency-test.js
@@ -8,8 +8,8 @@
 const EstimatedInputLatency = require('../../../computed/metrics/estimated-input-latency.js');
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
