+++ lighthouse-core/test/computed/metrics/interactive-test.js
@@ -9,8 +9,8 @@ const Interactive = require('../../../computed/metrics/interactive.js');
 
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 function generateNetworkRecords(records, navStart) {
   const navStartInMs = navStart / 1000;
