+++ lighthouse-core/test/computed/metrics/lantern-first-cpu-idle-test.js
@@ -8,8 +8,8 @@
 const LanternFirstCPUIdle = require('../../../computed/metrics/lantern-first-cpu-idle.js');
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
 describe('Metrics: Lantern TTFCPUI', () => {
