+++ lighthouse-core/test/computed/metrics/lantern-first-meaningful-paint-test.js
@@ -7,8 +7,8 @@
 
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 const LanternFirstMeaningfulPaint =
   require('../../../computed/metrics/lantern-first-meaningful-paint.js');
