+++ lighthouse-core/test/computed/metrics/lantern-interactive-test.js
@@ -8,8 +8,8 @@
 const LanternInteractive = require('../../../computed/metrics/lantern-interactive.js');
 const assert = require('assert');
 
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
 describe('Metrics: Lantern TTI', () => {
