+++ lighthouse-core/test/computed/metrics/max-potential-fid-test.js
@@ -8,8 +8,8 @@
 const assert = require('assert');
 
 const MaxPotentialFID = require('../../../computed/metrics/max-potential-fid.js');
-const trace = require('../../fixtures/traces/progressive-app-m60.json');
-const devtoolsLog = require('../../fixtures/traces/progressive-app-m60.devtools.log.json');
+const trace = require('../../fixtures/traces/progressive-app-m79.json');
+const devtoolsLog = require('../../fixtures/traces/progressive-app-m79.devtools.log.json');
 
 /* eslint-env jest */
